+++ src/index.ts
@@ -1,3 +1,4 @@
+import '@tensorflow/tfjs-backend-cpu';
 import '@tensorflow/tfjs-backend-webgl';
 import './utils';
 import './ui/css/styles.css';