+++ src/ui/components/TabList.svelte
@@ -1,0 +1,3 @@
+<div class="flex border-solid border-0 border-b border-gray-200 mb-2">
+  <slot />
+</div>