+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Marcelle changelog
 
+## 0.4.2
+
+- New `trainingHistory` component to track and monitor training runs.
+- `datasetTable` now automatically extracts columns from service data
+- Removed `onnxModel` from release due to limited operator support.
+- Improved documentation website
+
 ## 0.4.1
 
 - new component: `datasetTable` to display the contents of a dataset as a table